+++ models/intermediate/int_jira__issue_users.sql
@@ -10,7 +10,8 @@ with issue as (
         issue_id,
         coalesce(cast(revised_parent_issue_id as {{ dbt.type_string() }}), cast(parent_issue_id as {{ dbt.type_string() }})) as parent_issue_id,
 
-        {{ dbt_utils.star(from=ref('int_jira__issue_type_parents'), except=except_columns) }}
+        {{ dbt_utils.star(from=ref('int_jira__issue_type_parents'), 
+                            except=except_columns) }}
 
     from {{ ref('int_jira__issue_type_parents') }}
 ),