+++ models/intermediate/int_jira__issue_versions.sql
@@ -19,7 +19,10 @@ order_versions as (
     select
         *,
         -- using rank so batches stick together
-        rank() over (partition by field_id, issue_id order by updated_at desc) as row_num
+        rank() over (
+            partition by field_id, issue_id 
+            order by updated_at desc
+            ) as row_num
     from version_history
 ),