+++ models/jira__daily_sprint_issue_history.sql
@@ -2,11 +2,11 @@
     config(
         enabled=var('jira_using_sprints', True),
         materialized='incremental' if jira_is_incremental_compatible() else 'table',
-        partition_by = {'field': 'date_week', 'data_type': 'date'}
+        partition_by={'field': 'date_week', 'data_type': 'date'}
             if target.type not in ['spark', 'databricks'] else ['date_week'],
-        cluster_by = ['date_week'],
+        cluster_by=['date_week'],
         unique_key='sprint_issue_day_id',
-        incremental_strategy = 'insert_overwrite' if target.type in ('bigquery', 'databricks', 'spark') else 'delete+insert',
+        incremental_strategy='insert_overwrite' if target.type in ('bigquery', 'databricks', 'spark') else 'delete+insert',
         file_format='delta'
     )
 }}
@@ -22,33 +22,26 @@ with daily_issue_field_history as (
     {% endif %}
 ),
 
-issue_multiselect_history as (
-
-    select *
-    from {{ ref('int_jira__issue_multiselect_history') }}
-    {% if is_incremental() %}
-    where cast(updated_at as date) >= {{ max_date_week }}
-    {% endif %}
-),
-
-sprint as (
-
-    select *
-    from {{ var('sprint') }} 
-),
-
-issue as (
-
-    select *
-    from {{ var('issue') }} 
-),
-
-issue_sprint_history as (
-
+sprint_issue_pairing as (
     select
-        issue_multiselect_history.field_value as sprint_id,
-        issue_multiselect_history.issue_id,
-        issue_multiselect_history.updated_at as issue_assigned_to_sprint_at,
+        issue_id,
+        cast(field_value as {{ dbt.type_int() }}) as sprint_id,
+        updated_at,
+        is_active,
+        lead(field_value) over (partition by issue_id order by updated_at) as next_sprint,
+        lead(updated_at) over (partition by issue_id order by updated_at) as next_sprint_updated_at,
+        lag(is_active) over (partition by issue_id order by updated_at) as prev_is_active,
+        lag(updated_at) over (partition by issue_id order by updated_at) as prev_updated_at
+    from {{ ref('int_jira__issue_multiselect_history') }}
+    where field_name = 'sprint'
+        and field_value is not null
+),
+
+filtered_issue_sprint_history as (
+
+    select 
+        sprint_issue_pairing.*,
+        sprint.sprint_name,
         sprint.started_at as sprint_started_at,
         sprint.ended_at as sprint_ended_at,
         sprint.completed_at as sprint_completed_at,
@@ -59,54 +52,75 @@ issue_sprint_history as (
         issue.original_estimate_seconds,
         issue.remaining_estimate_seconds,
         issue.time_spent_seconds
-    from issue_multiselect_history
-    --Since we are only concerned with sprint data, thought it best to avoid issues not tied to sprints, hence the inner join.
-    inner join issue
-        on issue.issue_id = issue_multiselect_history.issue_id
-    inner join sprint
-        on issue_multiselect_history.field_value = cast(sprint.sprint_id as {{ dbt.type_string() }}) 
-    where issue_multiselect_history.field_name = 'sprint'
-        and issue_multiselect_history.is_active = true
-        and issue_multiselect_history.field_value is not null
+    from sprint_issue_pairing
+    inner join {{ var('issue') }} issue
+        on sprint_issue_pairing.issue_id = issue.issue_id
+    inner join {{ var('sprint') }} sprint
+        on sprint_issue_pairing.sprint_id = sprint.sprint_id
 ),
 
 issue_sprint_daily_history as (
-
     select
-        issue_sprint_history.sprint_id,
-        issue_sprint_history.issue_id,
+        filtered_issue_sprint_history.sprint_id,
+        filtered_issue_sprint_history.issue_id,
         daily_issue_field_history.date_day,
         daily_issue_field_history.date_week,
-        issue_sprint_history.issue_assigned_to_sprint_at,
-        issue_sprint_history.sprint_started_at,
-        issue_sprint_history.sprint_ended_at,
-        issue_sprint_history.sprint_completed_at,
-        issue_sprint_history.board_id,
-        issue_sprint_history.issue_created_at,
-        issue_sprint_history.issue_resolved_at,
-        issue_sprint_history.assignee_user_id,
-        issue_sprint_history.original_estimate_seconds,
-        issue_sprint_history.remaining_estimate_seconds,
-        issue_sprint_history.time_spent_seconds,
+        filtered_issue_sprint_history.updated_at as issue_assigned_to_sprint_at,
+        filtered_issue_sprint_history.sprint_name,
+        filtered_issue_sprint_history.sprint_started_at,
+        filtered_issue_sprint_history.sprint_ended_at,
+        filtered_issue_sprint_history.sprint_completed_at,
+        filtered_issue_sprint_history.board_id,
+        filtered_issue_sprint_history.issue_created_at,
+        filtered_issue_sprint_history.issue_resolved_at,
+        filtered_issue_sprint_history.assignee_user_id,
+        filtered_issue_sprint_history.original_estimate_seconds,
+        filtered_issue_sprint_history.remaining_estimate_seconds,
+        filtered_issue_sprint_history.time_spent_seconds,
         daily_issue_field_history.status,
-        daily_issue_field_history.status_id,
         cast(daily_issue_field_history.story_points as {{ dbt.type_float() }}) as story_points,
         cast(daily_issue_field_history.story_point_estimate as {{ dbt.type_float() }}) as story_point_estimate
-    from issue_sprint_history  
+    from filtered_issue_sprint_history  
     left join daily_issue_field_history 
-        on issue_sprint_history.issue_id = daily_issue_field_history.issue_id 
+        on filtered_issue_sprint_history.issue_id = daily_issue_field_history.issue_id
+        and cast(filtered_issue_sprint_history.updated_at as date) <= daily_issue_field_history.date_day
+        -- Ensure we always have a valid `date_day`
+        -- Ensure proper sprint transitions
+        and (
+            filtered_issue_sprint_history.is_active = true
+            or (
+                filtered_issue_sprint_history.is_active = false
+                and filtered_issue_sprint_history.next_sprint_updated_at is not null
+                and cast(filtered_issue_sprint_history.next_sprint_updated_at as date) > daily_issue_field_history.date_day
+            )
+        )
+    where daily_issue_field_history.date_day is not null
 ),
 
 issue_sprint_statuses as (
 
     select 
         issue_sprint_daily_history.*,
-        case when date_day >= cast(sprint_started_at as date) and date_day <= cast(sprint_ended_at as date) then true else false end as is_sprint_active,
-        case when date_day >= cast(sprint_completed_at as date) then true else false end as is_sprint_completed,
-        case when date_day >= cast(issue_created_at as date) and issue_resolved_at is null then true else false end as is_issue_open,
-        case when date_day >= cast(issue_resolved_at as date) and issue_resolved_at <= sprint_ended_at then true else false end as is_issue_resolved_in_sprint,
-        case when date_day >= cast(sprint_started_at as date) and date_day <= cast(sprint_ended_at as date)
-            then {{ dbt.datediff('date_day', 'sprint_ended_at', 'day') }} else null end as days_left_in_sprint
+        case when date_day >= cast(sprint_started_at as date) 
+            and (sprint_ended_at is null or date_day <= cast(sprint_ended_at as date)) 
+            then true else false 
+            end as is_sprint_active,
+        case when sprint_completed_at is not null 
+            and date_day >= cast(sprint_completed_at as date) 
+            then true else false 
+            end as is_sprint_completed,
+        case when date_day >= cast(issue_created_at as date) 
+            and issue_resolved_at is null 
+            then true else false 
+            end as is_issue_open,
+        case when date_day >= cast(issue_resolved_at as date) 
+            and issue_resolved_at <= sprint_ended_at 
+            then true else false 
+            end as is_issue_resolved_in_sprint,
+        case when date_day >= cast(sprint_started_at as date) 
+            and date_day <= cast(sprint_ended_at as date)
+            then {{ dbt.datediff('date_day', 'sprint_ended_at', 'day') }} else null 
+            end as days_left_in_sprint
     from issue_sprint_daily_history
 ),